streams: Refactor actions/subscriptions.py.
This is done to prevent a dependency cycle.

Fixes part of #21433.
