+++ zerver/actions/users.py
@@ -13,7 +13,7 @@ from django.utils.http import urlsafe_base64_encode
 from django.utils.timezone import now as timezone_now
 from django.utils.translation import get_language
 
-from zerver.actions.streams import send_peer_remove_events
+from zerver.actions.subscriptions import send_peer_remove_events
 from zerver.actions.user_groups import (
     do_send_user_group_members_update_event,
     update_users_in_full_members_system_group,
