+++ zerver/management/commands/add_users_to_streams.py
@@ -3,7 +3,7 @@ from typing import Any
 from django.core.management.base import CommandParser
 from typing_extensions import override
 
-from zerver.actions.streams import bulk_add_subscriptions
+from zerver.actions.subscriptions import bulk_add_subscriptions
 from zerver.lib.management import ZulipBaseCommand
 from zerver.lib.streams import ensure_stream
