+++ zerver/management/commands/remove_users_from_stream.py
@@ -3,7 +3,7 @@ from typing import Any
 from django.core.management.base import CommandParser
 from typing_extensions import override
 
-from zerver.actions.streams import bulk_remove_subscriptions
+from zerver.actions.subscriptions import bulk_remove_subscriptions
 from zerver.lib.management import ZulipBaseCommand
 from zerver.models.streams import get_stream
