+++ zerver/tests/test_event_queue.py
@@ -9,7 +9,7 @@ from typing_extensions import override
 
 from zerver.actions.message_send import internal_send_private_message
 from zerver.actions.muted_users import do_mute_user
-from zerver.actions.streams import do_change_subscription_property
+from zerver.actions.subscriptions import do_change_subscription_property
 from zerver.actions.user_groups import check_add_user_group
 from zerver.actions.user_settings import do_change_user_setting
 from zerver.actions.user_topics import do_set_user_topic_visibility_policy
