+++ zilencer/management/commands/add_mock_conversation.py
@@ -5,7 +5,7 @@ from typing_extensions import override
 from zerver.actions.create_user import do_create_user
 from zerver.actions.message_send import do_send_messages, internal_prep_stream_message
 from zerver.actions.reactions import do_add_reaction
-from zerver.actions.streams import bulk_add_subscriptions
+from zerver.actions.subscriptions import bulk_add_subscriptions
 from zerver.actions.user_settings import do_change_avatar_fields
 from zerver.lib.emoji import get_emoji_data
 from zerver.lib.management import ZulipBaseCommand
